refactor: CategoryPicker 위치 변경
