--- src/containers/home/ScheduleDrawer/inputs/CategoryPicker/index.ts
@@ -1,3 +1,0 @@
-import CategoryPicker from "./CategoryPicker";
-
-export default CategoryPicker;
