--- src/containers/home/ScheduleDrawer/inputs/CategoryPicker/pages/ExpenditureCategoryPage/index.ts
@@ -1,3 +1,0 @@
-import ExpenditureCategoryPage from "./ExpenditureCategoryPage";
-
-export default ExpenditureCategoryPage;
