--- src/containers/home/ScheduleDrawer/inputs/CategoryPicker/pages/IncomeCategoryPage/index.ts
@@ -1,3 +1,0 @@
-import IncomeCategoryPage from "./IncomeCategoryPage";
-
-export default IncomeCategoryPage;
