+++ src/containers/home/ScheduleDrawer/inputs/category-picker/CategoryPicker/CategoryPicker.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta } from "@storybook/react";
-import CategoryPicker from "./CategoryPicker";
+import CategoryPicker from "./CategoryPicker.tsx";
 
 // More on how to set up stories at: https://storybook.js.org/docs/react/writing-stories/introduction
 const meta = {
