+++ src/containers/home/ScheduleDrawer/inputs/category-picker/CategoryPicker/CategoryPicker.tsx
