+++ src/containers/home/ScheduleDrawer/inputs/category-picker/CategoryPicker/components/CategorySideTabButton.tsx
