+++ src/containers/home/ScheduleDrawer/inputs/category-picker/CategoryPicker/constants.ts
