+++ src/containers/home/ScheduleDrawer/inputs/category-picker/CategoryPicker/index.ts
@@ -1,0 +1,3 @@
+import CategoryPicker from "./CategoryPicker.tsx";
+
+export default CategoryPicker;
