+++ src/containers/home/ScheduleDrawer/inputs/category-picker/CategoryPicker/pages/ExpenditureCategoryPage/index.ts
@@ -1,0 +1,3 @@
+import ExpenditureCategoryPage from "./ExpenditureCategoryPage.tsx";
+
+export default ExpenditureCategoryPage;
