+++ src/containers/home/ScheduleDrawer/inputs/category-picker/CategoryPicker/pages/IncomeCategoryPage/IncomeCategoryPage.tsx
