+++ src/containers/home/ScheduleDrawer/inputs/category-picker/CategoryPicker/pages/IncomeCategoryPage/index.ts
@@ -1,0 +1,3 @@
+import IncomeCategoryPage from "./IncomeCategoryPage.tsx";
+
+export default IncomeCategoryPage;
